+++ js/state.js
@@ -1,8 +1,8 @@
+// Basically RAM for the app.
 import { createErrorModal } from "./utils";
 
 let state = {
   currentDocument: null,
-  unsavedDocuments: [],
 };
 
 // Gets the given state