+++ server/routes/export.js
@@ -1,8 +1,10 @@
+// Exporting TipTap documents as .pdf-files
+// Importing server functions
 const express = require("express");
 const router = express.Router();
 const fs = require("fs");
 const path = require("path");
-const puppeteer = require("puppeteer");
+const puppeteer = require("puppeteer"); // Puppeteer converts HTML into PDF.
 
 // Importing TipTap Extensions for Export
 const StarterKit = require("@tiptap/starter-kit").default;
@@ -12,6 +14,7 @@ const { renderToHTMLString } = require("@tiptap/static-renderer");
 const FileHandler = require("@tiptap/extension-file-handler").default;
 const Emoji = require("@tiptap/extension-emoji").default;
 
+// Configures all the TipTap extensions to be identical to editor.js
 async function setExtensions() {
   // Importing Extensions which cannot be imported (no commonjs version)
   const { default: CodeBlockLowlight } =
@@ -22,11 +25,12 @@ async function setExtensions() {
 
   return [
     StarterKit.configure({
-      codeBlock: false,
+      codeBlock: false, // Disabling codeBlock so that Syntax Highlighting works properly
     }),
     TableKit,
+    ListKit,
     Image.configure({
-      allowBase64: true,
+      inline: true,
       resize: {
         enabled: true,
         directions: ["top", "bottom", "left", "right"], // can be any direction or diagonal combination
@@ -36,13 +40,13 @@ async function setExtensions() {
       },
     }),
     FileHandler.configure({
+      allowedMimeTypes: ["image/png", "image/jpg", "image/gif"],
+      consumePasteEvent: true,
       onPaste: async (editor, files, htmlContent) => {
-        const base64 = await toBase64(files[0]);
-        editor.commands.setImage({ src: base64 });
-      },
-      onDrop: async (editor, files, pos) => {
-        const base64 = await toBase64(files[0]);
-        editor.commands.setImage({ src: base64 });
+        for (const file of files) {
+          const url = await uploadImage(file);
+          editor.chain().setImage({ src: url }).run();
+        }
       },
     }),
     Emoji,
@@ -54,9 +58,12 @@ async function setExtensions() {
   ];
 }
 
+// Processes HTML and converts it into PDF
 router.post("/api/export", async (req, res) => {
   const { exportDocument, name } = req.body;
 
+  // Basic HTML structure.
+  // The HTML from the request only includes the editor data, so stylesheets and stuff have to be set here.
   const exportHTML = `
   <html>
   <head>
@@ -71,21 +78,28 @@ router.post("/api/export", async (req, res) => {
   </html>
   `;
 
+  // Simulated browser which Puppeteer uses to generate the PDF.
   const browser = await puppeteer.launch({
     args: ["--allow-file-access-from-files"],
   });
+  // Creates new simulated page.
   const page = await browser.newPage();
+  // Sets the page content to the exportHTML
   const content = await page.setContent(exportHTML, {
+    // Waits until no network activities are active, ensuring that all async functions and similar
     waitUntil: "networkidle0",
   });
 
+  // Converts the page into a PDF with specifications.
   const pdf = await page.pdf({
     format: "A4",
     printBackground: true,
     margin: { top: "75px", bottom: "75px", left: "75px", right: "75px" },
   });
+  // Closes the simulated browser.
   await browser.close();
 
+  // Sends the PDF back to the client.
   res.setHeader("Content-Type", "application/pdf");
   res.setHeader(
     "Content-Disposition",