+++ server/server.js
@@ -7,6 +7,7 @@ const path = require("path");
 const app = express();
 const port = 8510;
 
+// Paths to Folders which need to exist within the data folder.
 const rootPath = path.join(__dirname, "../");
 const dataFolderPath = path.join(rootPath, "data");
 const autosavesFolderPath = path.join(dataFolderPath, "autosaves");
@@ -14,6 +15,19 @@ const notebooksFolderPath = path.join(dataFolderPath, "notebooks");
 const imageFolderPath = path.join(dataFolderPath, "images");
 const attachmentsFolderPath = path.join(dataFolderPath, "attachments");
 const masterFilePath = path.join(dataFolderPath, "master.json");
+
+// Server routes
+const documentsRoute = require("./routes/documents");
+const exportRoute = require("./routes/export");
+const autosaveRoute = require("./routes/autosave");
+const settingsRoute = require("./routes/settings");
+
+app.use(express.json());
+app.use(express.static(rootPath));
+app.use(documentsRoute);
+app.use(exportRoute);
+app.use(autosaveRoute);
+app.use(settingsRoute);
 
 const userDataFolders = [
   { name: "Data", path: dataFolderPath },
@@ -23,6 +37,7 @@ const userDataFolders = [
   { name: "Attachments", path: attachmentsFolderPath },
 ];
 
+// Creates any missing data folders.
 for (let i = 0; i < userDataFolders.length; i++) {
   if (!fs.existsSync(userDataFolders[i].path)) {
     fs.mkdirSync(userDataFolders[i].path);
@@ -33,11 +48,13 @@ for (let i = 0; i < userDataFolders.length; i++) {
   }
 }
 
-// Master File for config across Sessions
+// Masterfile to store config across Sessions
 if (!fs.existsSync(masterFilePath)) {
   const masterFileContent = `[
   {
-    "unsavedFiles": []
+    "unsavedFiles": [],
+    "autosaveInterval": 10,
+    "helpTextHoverTime": 15
   }
 ]
 `;
@@ -47,37 +64,92 @@ if (!fs.existsSync(masterFilePath)) {
     "This is standard if you've freshly cloned the Repository, as the data folder is ignored by git.",
   );
 } else {
+  // If the masterfile is found, it's checked for any deprecated or missing properties, making updates to it automatic.
   deleteDeprecatedMasterProperties();
+  addMissingMasterProperties();
+}
+
+// Updates the masterfile and gives feedback on success.
+// This function is used after deprecated / missing properties are found.
+async function updateMasterfile(masterFile) {
+  try {
+    fs.writeFileSync(masterFilePath, JSON.stringify(masterFile), "utf-8");
+    return true;
+  } catch (err) {
+    console.error("Something went wrong trying to update master.json.");
+    console.error(err);
+    return false;
+  }
 }
 
 // Deletes deprecated master.json properties
 async function deleteDeprecatedMasterProperties() {
   console.log("Checking for deprecated master.json properties...");
 
-  let changesMade = false;
   // Getting the masterfile data. Has to be parsed.
   const rawMasterFile = await fs.readFileSync(masterFilePath, "utf-8");
   const masterFile = JSON.parse(rawMasterFile);
+  let changesMade = false;
+  // Array of ever property to be released and later be deprecated.
+  const deprecatedProperties = ["usedImages"];
 
-  // Deletes the usedImages property
-  if (masterFile[0].usedImages) {
-    delete masterFile[0].usedImages;
-    console.log('Deleted masterfile property "usedImages".');
+  // Deletes deprecated properties
+  for (let i = 0; i < deprecatedProperties.length; i++) {
+    if (masterFile[0][deprecatedProperties[i]]) {
+      delete masterFile[0][deprecatedProperties[i]];
+      console.log(`Deleted masterfile property "${deprecatedProperties[i]}".`);
+      changesMade = true;
+    }
+  }
+
+  // Updates the masterfile if changes were made
+  if (changesMade) {
+    if (updateMasterfile(masterFile)) {
+      console.log("Successfully removed deprecated masterfile properties.");
+    } else {
+      console.error(
+        "Something went wrong trying to remove deprecated masterfile properties.",
+      );
+    }
+  } else {
+    console.log("No deprecated masterfile properties found.");
+  }
+}
+
+// Adds any missing properties to the master.json file.
+async function addMissingMasterProperties() {
+  console.log("Checking for missing master.json properties...");
+
+  // Getting the masterfile data. Has to be parsed.
+  const rawMasterFile = await fs.readFileSync(masterFilePath, "utf-8");
+  const masterFile = JSON.parse(rawMasterFile);
+  let changesMade = false;
+  // Every property which should be in the masterfile.
+  const allProperties = {
+    unsavedFiles: [],
+    autosaveInterval: 10000,
+    helpTextHoverTime: 15,
+  };
+
+  // Adds all the missing properties
+  for (const key in allProperties) {
+    if (!masterFile[0][key]) {
+      masterFile[0][key] = allProperties[key];
+      console.log(`Added missing masterfile property "${key}".`);
+    }
     changesMade = true;
   }
 
+  // Updates the masterfile if changes were made
   if (changesMade) {
-    try {
-      // Updates the masterfile
-      fs.writeFileSync(masterFilePath, JSON.stringify(masterFile), "utf-8");
-      console.log("Successfully removed deprecated masterfile properties.");
-    } catch (err) {
-      console.error("Something went wrong trying to update master.json.");
-      console.error(err);
+    if (updateMasterfile(masterFile)) {
+      console.log("Successfully added missing masterfile properties.");
+    } else {
+      console.log("Failed to add missing masterfile properties.");
     }
   } else {
     console.log(
-      "No deprecated masterfile properties found. Starting App without changes.",
+      "No missing masterfile properties found. Starting App without changes.",
     );
   }
 }
@@ -94,16 +166,7 @@ app.get("/api/getMaster", async (req, res) => {
   }
 });
 
-const documentsRoute = require("./routes/documents");
-const exportRoute = require("./routes/export");
-const versionsRoute = require("./routes/versions");
-
-app.use(express.json());
-app.use(express.static(rootPath));
-app.use(documentsRoute);
-app.use(exportRoute);
-app.use(versionsRoute);
-
+// Sends index.html to the client.
 app.get("/", (req, res) => {
   res.sendFile(path.join(rootPath, "index.html"));
 });